feat/CIV-9127 Refactor to use util's pageSectionBuilder
