+++ src/main/common/models/caseProgression/uploadYourDocumentsSectionBuilder.ts
@@ -3,30 +3,6 @@ import {PageSectionBuilder} from 'common/utils/pageSectionBuilder';
 import {t} from 'i18next';
 export class UploadYourDocumentsSectionBuilder extends PageSectionBuilder {
   _claimSummarySections: ClaimSummarySection[] = [];
-  addMainTitle(mainTitle: string, variables?: unknown) {
-    const mainTitleSection = ({
-      type: ClaimSummaryType.MAINTITLE,
-      data: {
-        text: mainTitle,
-        variables: variables,
-      },
-    });
-    this._claimSummarySections.push(mainTitleSection);
-    return this;
-  }
-
-  addLeadParagraph(text: string, variables?: unknown, classes?: string) {
-    const leadParagraphSection = ({
-      type: ClaimSummaryType.LEAD_PARAGRAPH,
-      data: {
-        text: text,
-        variables: variables,
-        classes: classes,
-      },
-    });
-    this._claimSummarySections.push(leadParagraphSection);
-    return this;
-  }
 
   addStartButton(title: string, href: string) {
     const startButtonSection = ({
