+++ src/main/common/utils/pageSectionBuilder.ts
@@ -55,6 +55,31 @@ export class PageSectionBuilder {
     return this;
   }
 
+  addMainTitle(mainTitle: string, variables?: unknown) {
+    const mainTitleSection = ({
+      type: ClaimSummaryType.MAINTITLE,
+      data: {
+        text: mainTitle,
+        variables: variables,
+      },
+    });
+    this._claimSummarySections.push(mainTitleSection);
+    return this;
+  }
+
+  addLeadParagraph(text: string, variables?: unknown, classes?: string) {
+    const leadParagraphSection = ({
+      type: ClaimSummaryType.LEAD_PARAGRAPH,
+      data: {
+        text: text,
+        variables: variables,
+        classes: classes,
+      },
+    });
+    this._claimSummarySections.push(leadParagraphSection);
+    return this;
+  }
+
   build() {
     return this._claimSummarySections;
   }
