+++ src/test/unit/common/models/caseProgression/uploadYourDocumentsSectionBuilder.test.ts
@@ -3,45 +3,6 @@ import {ClaimSummaryType} from 'form/models/claimSummarySection';
 import {t} from 'i18next';
 
 describe('UploadYourDocumentsSectionBuilder tests', ()=> {
-  it('should create mainTitle', ()=> {
-  //Given
-    const mainTitleExpected = ({
-      type: ClaimSummaryType.MAINTITLE,
-      data: {
-        text: 'text',
-        variables: 'variables',
-      },
-    });
-
-    //When
-    const mainTitleBuilt = new UploadYourDocumentsSectionBuilder()
-      .addMainTitle(mainTitleExpected.data.text,mainTitleExpected.data.variables)
-      .build();
-
-    //Then
-    expect(mainTitleBuilt).toEqual([mainTitleExpected]);
-  });
-
-  it('should add leadParagraph', ()=> {
-    //Given
-    const leadParagraphExpected = ({
-      type: ClaimSummaryType.LEAD_PARAGRAPH,
-      data: {
-        text: 'text',
-        variables: 'variables',
-        classes: 'classes',
-      },
-    });
-
-    //When
-    const leadParagraphBuilt = new UploadYourDocumentsSectionBuilder()
-      .addLeadParagraph(leadParagraphExpected.data.text,leadParagraphExpected.data.variables,leadParagraphExpected.data.classes)
-      .build();
-
-    //Then
-    expect(leadParagraphBuilt).toEqual([leadParagraphExpected]);
-  });
-
   it('should add Inset Text', ()=> {
     //Given
     const insetTextExpected = ({
