+++ src/test/unit/common/utils/pageSectionBuilder.test.ts
@@ -153,4 +153,43 @@ describe('PageSectionBuilder tests', ()=> {
     //Then
     expect(buttonBuilt).toEqual([buttonExpected]);
   });
+
+  it('should create mainTitle', ()=> {
+    //Given
+    const mainTitleExpected = ({
+      type: ClaimSummaryType.MAINTITLE,
+      data: {
+        text: 'text',
+        variables: 'variables',
+      },
+    });
+
+    //When
+    const mainTitleBuilt = new PageSectionBuilder()
+      .addMainTitle(mainTitleExpected.data.text,mainTitleExpected.data.variables)
+      .build();
+
+    //Then
+    expect(mainTitleBuilt).toEqual([mainTitleExpected]);
+  });
+
+  it('should add leadParagraph', ()=> {
+    //Given
+    const leadParagraphExpected = ({
+      type: ClaimSummaryType.LEAD_PARAGRAPH,
+      data: {
+        text: 'text',
+        variables: 'variables',
+        classes: 'classes',
+      },
+    });
+
+    //When
+    const leadParagraphBuilt = new PageSectionBuilder()
+      .addLeadParagraph(leadParagraphExpected.data.text,leadParagraphExpected.data.variables,leadParagraphExpected.data.classes)
+      .build();
+
+    //Then
+    expect(leadParagraphBuilt).toEqual([leadParagraphExpected]);
+  });
 });
